260107 라이더 신청 시 주소 좌표(lat,lng) 저장 누락 문제 해결 - rider.field.js : 2-50글자 -> 2-100글자 변경 - Rider.js : BIGINT -> DECIMAL 수정 - 08dummy-images.seeder.js 추가

+++ app/models/Rider.js
@@ -2,6 +2,7 @@
  * @file app/models/Rider.js
  * @description rider model
  * 251217 v1.0.0 wook 초기 생성
+ * 260107 sara update BIGINT에서 수정
  */
 
 import { DataTypes } from 'sequelize';
@@ -78,13 +79,13 @@ const attributes = {
   },
   lat: {
     field: 'lat',
-    type: DataTypes.BIGINT,
+    type: DataTypes.DECIMAL(10, 8), // 260107 sara update BIGINT에서 수정
     allowNull: true,
     comment: '위도'
   },
   lng: {
     field: 'lng',
-    type: DataTypes.BIGINT,
+    type: DataTypes.DECIMAL(11, 8), // 260107 sara update BIGINT에서 수정
     allowNull: true,
     comment: '경도'
   },
